add specification annotations
